add index and schema por Navigation breadcrumb
