+++ src/components/FsSlider.jsx
@@ -1,4 +1,3 @@
-import React, { useState } from "react";
 // swiper components
 import { Swiper, SwiperSlide } from "swiper/react";
 // swiper styles
@@ -9,8 +8,8 @@ import "swiper/css/navigation";
 import { Pagination, Navigation } from "swiper/modules";
 
 // components
-import NavigationBreadcrumb from "./NavigationBreadcrumb.jsx";
-import SkillsIcons from "./SkillsIcons.jsx";
+import NavigationBreadcrumb from "./NavigationBreadcrumb/NavigationBreadcrumb.jsx";
+import { SkillsIcons } from "./SkillsIcons";
 // data
 import { FullstackData } from "../data/FsProjects.jsx";
 import { FsSkills } from "../data/FsSkills.jsx";
@@ -19,19 +18,10 @@ import { FsSkills } from "../data/FsSkills.jsx";
 import { LazyLoadImage } from "react-lazy-load-image-component";
 
 const FsSlider = () => {
-    const [videoUrl, setVideoUrl] = useState(null);
     const fsData = FullstackData;
 
     const handleClick = (url) => {
         window.open(url, "_blank");
-    };
-
-    const handleVideoClick = (url) => {
-        setVideoUrl(url);
-    };
-
-    const closeVideoPlayer = () => {
-        setVideoUrl(null);
     };
 
     return (
@@ -53,7 +43,7 @@ const FsSlider = () => {
                     ]}></NavigationBreadcrumb>
             </section>
             <section className="row skills_slider">
-                <section className="col-5 col-md-5 col-lg-5 col-sm-12 flex_col ps-4">
+                <section className="col-5 col-md-5 col-lg-5 flex_col">
                     <SkillsIcons skillsData={FsSkills} />
                 </section>
                 <Swiper
@@ -94,18 +84,6 @@ const FsSlider = () => {
                                             <span>Visita esta página</span>
                                         </button>
                                     )}
-                                    {fsProject.video_src && (
-                                        <button
-                                            onClick={() =>
-                                                handleVideoClick(
-                                                    fsProject.video_src
-                                                )
-                                            }
-                                            className="btn btn-secondary cursor-pointer flex_col me-4"
-                                            aria-label="Ver video">
-                                            <span>Ver video</span>
-                                        </button>
-                                    )}
                                     <button
                                         onClick={() =>
                                             handleClick(fsProject.rep_url)
@@ -120,32 +98,6 @@ const FsSlider = () => {
                     ))}
                 </Swiper>
             </section>
-
-            {/* Video Player */}
-            {videoUrl && (
-                <section
-                    className="video-player-overlay"
-                    onClick={closeVideoPlayer}>
-                    <article
-                        className="video-player"
-                        onClick={(e) => e.stopPropagation()}>
-                        <button
-                            className="close-button"
-                            onClick={closeVideoPlayer}>
-                            &times;
-                        </button>
-                        <video
-                            src={videoUrl}
-                            controls
-                            controlsList="nodownload noremoteplayback"
-                            disablePictureInPicture
-                            onContextMenu={(e) => e.preventDefault()}
-                            className="video-element">
-                            Tu navegador no soporta la reproducción de video.
-                        </video>
-                    </article>
-                </section>
-            )}
         </>
     );
 };
