+++ src/components/NavigationBreadcrumb/index.js
@@ -1,0 +1,1 @@
+export * from "./NavigationBreadcrumb";
